+++ app/tarot-read/loading.tsx
@@ -1,5 +1,0 @@
-import { LoadingPage } from '../../components/loadingPage/LoadingPage';
-
-export default function Loading() {
-  return <LoadingPage />;
-}